fix: moved all to publication service
