+++ server/src/main/java/com/epam/aidial/core/server/service/PublicationService.java
@@ -2,6 +2,7 @@ package com.epam.aidial.core.server.service;
 
 import com.epam.aidial.core.config.Application;
 import com.epam.aidial.core.server.ProxyContext;
+import com.epam.aidial.core.server.config.ConfigStore;
 import com.epam.aidial.core.server.controller.ApplicationUtil;
 import com.epam.aidial.core.server.data.ListPublishedResourcesRequest;
 import com.epam.aidial.core.server.data.Notification;
@@ -12,25 +13,31 @@ import com.epam.aidial.core.server.data.ResourceUrl;
 import com.epam.aidial.core.server.data.Rule;
 import com.epam.aidial.core.server.security.AccessService;
 import com.epam.aidial.core.server.security.EncryptionService;
-import com.epam.aidial.core.server.service.schemarichapps.ApplicationResourcesReplacementService;
-import com.epam.aidial.core.server.service.schemarichapps.PublicationEnrichmentService;
+import com.epam.aidial.core.server.util.ApplicationTypeSchemaUtils;
 import com.epam.aidial.core.server.util.BucketBuilder;
 import com.epam.aidial.core.server.util.ProxyUtil;
 import com.epam.aidial.core.server.util.ResourceDescriptorFactory;
 import com.epam.aidial.core.storage.data.MetadataBase;
+import com.epam.aidial.core.storage.data.NodeType;
 import com.epam.aidial.core.storage.data.ResourceFolderMetadata;
 import com.epam.aidial.core.storage.data.ResourceItemMetadata;
 import com.epam.aidial.core.storage.data.UserMetadata;
+import com.epam.aidial.core.storage.http.HttpException;
+import com.epam.aidial.core.storage.http.HttpStatus;
 import com.epam.aidial.core.storage.resource.ResourceDescriptor;
 import com.epam.aidial.core.storage.resource.ResourceType;
 import com.epam.aidial.core.storage.service.ResourceService;
 import com.epam.aidial.core.storage.util.EtagHeader;
 import com.epam.aidial.core.storage.util.UrlUtil;
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.mutable.MutableObject;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -41,8 +48,8 @@ import java.util.Set;
 import java.util.function.LongSupplier;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import javax.annotation.Nullable;
-
 
 @Slf4j
 @RequiredArgsConstructor
@@ -59,7 +66,6 @@ public class PublicationService {
     private static final Set<ResourceType> ALLOWED_RESOURCES = Set.of(ResourceTypes.FILE, ResourceTypes.CONVERSATION,
             ResourceTypes.PROMPT, ResourceTypes.APPLICATION);
 
-    private final PublicationEnrichmentService publicationEnrichmentService;
     private final EncryptionService encryption;
     private final ResourceService resourceService;
     private final AccessService accessService;
@@ -69,7 +75,7 @@ public class PublicationService {
     private final ResourceOperationService resourceOperationService;
     private final Supplier<String> ids;
     private final LongSupplier clock;
-    private final ApplicationResourcesReplacementService applicationResourcesReplacementService;
+    private final ConfigStore configStore;
 
     public static boolean isReviewBucket(ResourceDescriptor resource) {
         return resource.isPrivate() && resource.getBucketLocation().contains(PUBLICATIONS_NAME);
@@ -352,7 +358,7 @@ public class PublicationService {
         publication.setStatus(Publication.Status.PENDING);
         publication.setAuthor(context.getUserDisplayName());
 
-        publicationEnrichmentService.enrichPublicationWithSchemaRichApplicationFiles(publication);
+        addSchemaRichApplicationFiles(publication);
 
         Set<String> urls = new HashSet<>();
         for (Publication.Resource resource : publication.getResources()) {
@@ -550,7 +556,7 @@ public class PublicationService {
 
             if (from.getType() == ResourceTypes.APPLICATION) {
                 applicationService.copyApplication(from, to, null, false, app -> {
-                    applicationResourcesReplacementService.replaceSchemaRichAppOwnResources(app, to.getUrl(), replacementLinks);
+                    replaceSchemaRichAppOwnResources(app, to.getUrl(), replacementLinks);
                     app.setReference(ApplicationUtil.generateReference());
                     app.setIconUrl(replaceLink(replacementLinks, app.getIconUrl()));
                 });
@@ -592,7 +598,7 @@ public class PublicationService {
 
             if (from.getType() == ResourceTypes.APPLICATION) {
                 applicationService.copyApplication(from, to, publication.getDisplayAuthor(), false, app -> {
-                    applicationResourcesReplacementService.replaceSchemaRichAppOwnResources(app, to.getUrl(), replacementLinks);
+                    replaceSchemaRichAppOwnResources(app, to.getUrl(), replacementLinks);
                     app.setReference(ApplicationUtil.generateReference());
                     app.setIconUrl(replaceLink(replacementLinks, app.getIconUrl()));
                 });
@@ -702,5 +708,258 @@ public class PublicationService {
         }
 
         return url;
+    }
+
+    public void addSchemaRichApplicationFiles(Publication publication) {
+        if (publication.getResources().isEmpty()) {
+            return;
+        }
+
+        List<String> sourceUrlsFromRequest = publication.getResources().stream()
+                .map(Publication.Resource::getSourceUrl)
+                .toList();
+
+        Map<String, Integer> fileNamesTaken = new HashMap<>();
+
+        List<Publication.Resource> newResources = publication.getResources().stream()
+                .filter(resource -> resource.getAction() != Publication.ResourceAction.DELETE)
+                .filter(this::isApplicationResource)
+                .flatMap(resource -> getApplicationFilesAndFolders(resource, sourceUrlsFromRequest, fileNamesTaken))
+                .toList();
+
+        // Check for duplicates in newResources (overlap folder's content with files alone)
+        Map<String, Publication.Resource> sourceUrlMap = new HashMap<>();
+        for (Publication.Resource resource : newResources) {
+            String sourceUrl = resource.getSourceUrl();
+            if (sourceUrlMap.containsKey(sourceUrl)) {
+                throw new HttpException(HttpStatus.BAD_REQUEST, "Duplicate source URL found: " + sourceUrl);
+            }
+            sourceUrlMap.put(sourceUrl, resource);
+        }
+
+        publication.getResources().addAll(newResources);
+    }
+
+    private boolean isApplicationResource(Publication.Resource resource) {
+        ResourceDescriptor resourceDescriptor = ResourceDescriptorFactory.fromAnyUrl(resource.getSourceUrl(), encryption);
+        return !resourceDescriptor.isFolder() && resourceDescriptor.getType() == ResourceTypes.APPLICATION;
+    }
+
+    private Stream<Publication.Resource> getApplicationFilesAndFolders(Publication.Resource pubicationResource,
+                                                                       List<String> otherSourceUrlsFromRequest, Map<String, Integer> fileNamesTaken) {
+        ResourceDescriptor resourceToPublish = ResourceDescriptorFactory.fromAnyUrl(pubicationResource.getSourceUrl(), encryption);
+        if (resourceToPublish.getType() != ResourceTypes.APPLICATION) {
+            return Stream.empty();
+        }
+
+        Application applicationToPublish = applicationService.getApplication(resourceToPublish).getValue();
+        if (applicationToPublish.getApplicationTypeSchemaId() == null) {
+            return Stream.empty();
+        }
+
+        String targetFolder = getTargetFolderForCustomAppFiles(pubicationResource.getTargetUrl(), encryption);
+
+        List<ResourceDescriptor> applicationsOwnDescriptors = ApplicationTypeSchemaUtils.getFiles(configStore.get(), applicationToPublish, encryption, resourceService)
+                .stream()
+                .filter(descriptor -> !otherSourceUrlsFromRequest.contains(descriptor.getUrl()))
+                .toList();
+
+        Stream<Publication.Resource> folderDescriptors = applicationsOwnDescriptors.stream()
+                .filter(ResourceDescriptor::isFolder)
+                .flatMap(folder -> createResourcesForFolderFiles(folder, targetFolder, fileNamesTaken, pubicationResource.getAction()));
+
+        Stream<Publication.Resource> fileDescriptors = applicationsOwnDescriptors.stream()
+                .filter(descriptor -> !descriptor.isFolder())
+                .map(file -> createResourceForStandaloneFile(file, targetFolder, fileNamesTaken, pubicationResource.getAction()));
+
+        return Stream.concat(folderDescriptors, fileDescriptors);
+    }
+
+    private Stream<Publication.Resource> createResourcesForFolderFiles(ResourceDescriptor sourceFolderDescriptor, String targetFolderUrl,
+                                                                       Map<String, Integer> fileNamesTaken, Publication.ResourceAction action) {
+        String targetSubFolderUrl = createUniqueTargetResourceUrl(sourceFolderDescriptor, targetFolderUrl, fileNamesTaken);
+        return listPrivateFilesFromFolderWithSubFolders(sourceFolderDescriptor)
+                .map(sourceFileDescriptor ->
+                        createResourceForFileInFolder(
+                                sourceFileDescriptor, sourceFolderDescriptor, targetSubFolderUrl, action)
+                );
+    }
+
+    private static Publication.Resource createResourceForStandaloneFile(ResourceDescriptor sourceFileDescriptor,
+                                                                        String targetFolderUrl,
+                                                                        Map<String, Integer> fileNamesTaken,
+                                                                        Publication.ResourceAction action) {
+        String targetUrl = createUniqueTargetResourceUrl(sourceFileDescriptor, targetFolderUrl, fileNamesTaken);
+
+        return new Publication.Resource()
+                .setAction(action)
+                .setSourceUrl(sourceFileDescriptor.getUrl())
+                .setTargetUrl(targetUrl);
+    }
+
+    private static String createUniqueTargetResourceUrl(ResourceDescriptor sourceDescriptor, String targetFolderUrl, Map<String, Integer> fileNamesTaken) {
+        String fileName = sourceDescriptor.getName();
+        int count = fileNamesTaken.getOrDefault(fileName, 0) + 1;
+        fileNamesTaken.put(fileName, count);
+
+        if (count > 1) {
+            // Add counter to filename while preserving extension
+            if (fileName.contains(".")) {
+                // File has extension
+                fileName = fileName.replaceFirst("(\\.[^.]+)$", "_" + count + "$1");
+            } else {
+                // File has no extension
+                fileName = fileName + "_" + count;
+            }
+        }
+
+        return ResourceDescriptorFactory.fromDecoded(
+                sourceDescriptor.getType(),
+                ResourceDescriptor.PUBLIC_BUCKET,
+                ResourceDescriptor.PATH_SEPARATOR,
+                targetFolderUrl + fileName).getUrl();
+    }
+
+    private static Publication.Resource createResourceForFileInFolder(ResourceDescriptor sourceFileDescriptor,
+                                                                      ResourceDescriptor sourceFolderDescriptor,
+                                                                      String targetFolderUrl,
+                                                                      Publication.ResourceAction action) {
+        String relativeFilePath = sourceFolderDescriptor.getRelativePath(sourceFileDescriptor);
+        String targetUrl = targetFolderUrl + ResourceDescriptor.PATH_SEPARATOR + relativeFilePath;
+        return new Publication.Resource()
+                .setAction(action)
+                .setSourceUrl(sourceFileDescriptor.getUrl())
+                .setTargetUrl(targetUrl);
+    }
+
+    public Stream<ResourceDescriptor> listPrivateFilesFromFolderWithSubFolders(ResourceDescriptor folderDescriptor) {
+        if (!folderDescriptor.isFolder()) {
+            return Stream.empty();
+        }
+
+        List<ResourceDescriptor> fileDescriptors = new ArrayList<>();
+        String nextToken = null;
+
+        do {
+            try {
+                ResourceFolderMetadata folderMetadata =
+                        resourceService.getFolderMetadata(folderDescriptor, nextToken, 1000, true);
+
+                if (folderMetadata == null || folderMetadata.getItems() == null) {
+                    break;
+                }
+
+                // Process all files in this page
+                folderMetadata.getItems().stream()
+                        .filter(item -> item.getNodeType() != NodeType.FOLDER)
+                        .map(item -> ResourceDescriptorFactory.fromPrivateUrl(item.getUrl(), encryption))
+                        .forEach(fileDescriptors::add);
+
+                nextToken = folderMetadata.getNextToken();
+            } catch (Exception e) {
+                log.warn("Failed to list files in folder while publishing: {}", folderDescriptor.getUrl(), e);
+                throw new RuntimeException(e);
+            }
+        } while (nextToken != null);
+
+        return fileDescriptors.stream();
+    }
+
+    public void replaceSchemaRichAppOwnResources(Application application, String targetApplicationUrl, Map<String, String> replacementLinks) {
+        if (application.getApplicationTypeSchemaId() == null) {
+            return;
+        }
+        String targetApplicationResourceFolderUrl = getTargetFolderForCustomAppFiles(targetApplicationUrl, encryption);
+        replacementLinks = extractApplicationOwnResourcesMapping(application, targetApplicationResourceFolderUrl, replacementLinks);
+        applyApplicationOwnResourcesMapping(application, replacementLinks);
+    }
+
+    private Map<String, String> extractApplicationOwnResourcesMapping(Application application, String targetApplicationResourceFolderUrl, Map<String, String> replacementLinks) {
+        List<ResourceDescriptor> applicationOwnResources = ApplicationTypeSchemaUtils.getFiles(
+                configStore.get(), application, encryption, resourceService);
+
+        Map<String, String> resultMapping = new HashMap<>();
+
+        for (ResourceDescriptor resource : applicationOwnResources) {
+            String resourceUrl = resource.getUrl();
+            if (resource.isFolder()) {
+                String folderReplacement = replacementLinks.entrySet().stream()
+                        .filter(entry -> entry.getKey().startsWith(resourceUrl))
+                        .map(Map.Entry::getValue)
+                        .findFirst()
+                        .orElseThrow(() -> new IllegalStateException("Missing replacement link for folder: " + resourceUrl));
+                resultMapping.put(resourceUrl, extractFirstPathComponent(targetApplicationResourceFolderUrl, folderReplacement) + ResourceDescriptor.PATH_SEPARATOR);
+            } else {
+                String fileReplacement = replacementLinks.get(resourceUrl);
+                if (fileReplacement == null) {
+                    throw new IllegalStateException("Missing replacement link for file: " + resourceUrl);
+                }
+                resultMapping.put(resourceUrl, extractFirstPathComponent(targetApplicationResourceFolderUrl, fileReplacement));
+            }
+        }
+
+        return resultMapping;
+    }
+
+    private static String extractFirstPathComponent(String basePath, String fullPath) {
+        int basePathIndex = fullPath.indexOf(basePath);
+        if (basePathIndex == -1) {
+            throw new IllegalStateException(
+                    "Base path '" + basePath + "' not found in full path '" + fullPath + "'");
+        }
+
+        String prefixPath = fullPath.substring(0, basePathIndex);
+        String relativePath = fullPath.substring(basePathIndex + basePath.length());
+        String firstComponent = relativePath.split(ResourceDescriptor.PATH_SEPARATOR, 2)[0];
+
+        return prefixPath + basePath + firstComponent;
+    }
+
+    private static void applyApplicationOwnResourcesMapping(Application application, Map<String, String> replacementLinks) {
+        JsonNode customProperties = ProxyUtil.MAPPER.convertValue(application.getApplicationProperties(), JsonNode.class);
+        replaceTextNodes(customProperties, replacementLinks, null, null);
+        Map<String, Object> customPropertiesMap = ProxyUtil.MAPPER.convertValue(customProperties, new TypeReference<>() {
+        });
+        application.setApplicationProperties(customPropertiesMap);
+    }
+
+    private static void replaceTextNodes(JsonNode node, Map<String, String> replacementMap, JsonNode parent, String fieldName) {
+        if (node.isObject()) {
+            node.fields().forEachRemaining(entry -> replaceTextNodes(entry.getValue(), replacementMap, node, entry.getKey()));
+        } else if (node.isArray()) {
+            for (int i = 0; i < node.size(); i++) {
+                JsonNode childNode = node.get(i);
+                if (childNode.isTextual()) {
+                    String replacement = replacementMap.get(childNode.textValue());
+                    if (replacement != null) {
+                        ((ArrayNode) node).set(i, replacement);
+                    }
+                } else {
+                    replaceTextNodes(childNode, replacementMap, node, String.valueOf(i));
+                }
+            }
+        } else if (node.isTextual()) {
+            String replacement = replacementMap.get(node.textValue());
+            if (replacement != null && parent.isObject()) {
+                ((ObjectNode) parent).put(fieldName, replacement);
+            }
+        }
+    }
+
+    public static String getTargetFolderForCustomAppFiles(String targetUrl, EncryptionService encryptionService) {
+        ResourceDescriptor targetResourceDescriptor = ResourceDescriptorFactory.fromAnyUrl(targetUrl, encryptionService);
+        if (targetResourceDescriptor.isFolder()) {
+            throw new IllegalArgumentException("Target url must be a file");
+        }
+        if (targetResourceDescriptor.getType() != ResourceTypes.APPLICATION) {
+            throw new IllegalArgumentException("Target url must be an application type");
+        }
+        String appName = targetResourceDescriptor.getName();
+        String appPath = targetResourceDescriptor.getParentPath();
+        if (appPath == null) {
+            return "." + appName + ResourceDescriptor.PATH_SEPARATOR;
+        } else {
+            return appPath + ResourceDescriptor.PATH_SEPARATOR + "." + appName + ResourceDescriptor.PATH_SEPARATOR;
+        }
     }
 }
